+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add details on using schema backend directly in README.
 - Default to the provided schema if topic is not provided when
   encoding to `AvroSchemaRegistry`.
+- Add mapping syntax for the `schema` call in `SchemaControllerMixin`.
 
 ## 1.8.2-beta1 - 2020-09-09